pyproject.toml
[tool.poetry]
name = "backend_django"
version = "0.1.0"
description = ""
authors = ["Valexandrin <alex-vadim@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
mypy = "0.942"
flake8 = "^4.0.1"
wemake-python-styleguide = "^0.16.1"
pydocstyle = "^6.1.1"
django-stubs = {extras = ["compatible-mypy"], version = "^1.11.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
